added secret to authenticate with webhook request

+++ src/pages/api/activities/webhook.ts
@@ -1,5 +1,6 @@
 import { getFirestore } from "firebase-admin/firestore";
 import { getMessaging } from "firebase-admin/messaging";
+import { StatusCodes } from "http-status-codes";
 import { NextApiRequest, NextApiResponse } from "next";
 
 import { getFirebaseAdminApp } from "../utils/firebase-admin";
@@ -12,6 +13,11 @@ export default async function handler(
 
   if (req.method === "POST") {
     const body: Activity = req.body;
+    const { secret } = req.query;
+
+    if (!secret || secret !== process.env.WEBHOOK_SECRET)
+      return res.status(StatusCodes.UNAUTHORIZED).json({});
+
     const messaging = getMessaging(firebaseAdminApp);
     const db = getFirestore(firebaseAdminApp);
     const ref = db.collection(`fcm_tokens/users/${body.to_user.id}`);
@@ -20,7 +26,10 @@ export default async function handler(
 
     const result = await messaging.sendMulticast({
       tokens,
-      notification: { title: "hello world", body: "Cool beans" },
+      notification: {
+        title: "New follow",
+        body: `You have a new follower: ${body.to_user.username}`,
+      },
     });
 
     if (result.failureCount > 0) {
@@ -28,16 +37,14 @@ export default async function handler(
         if (!result.responses[i].success) {
           const document = (await ref.where("fcm_token", "==", tokens[i]).get())
             .docs[0];
-          if (document) {
-            await document.ref.delete();
-          }
+          if (document) await document.ref.delete();
         }
       }
     }
-    return res.status(200).json({});
+    return res.status(StatusCodes.OK).json({});
   }
 
-  return res.status(404).json({});
+  return res.status(StatusCodes.NOT_FOUND).json({});
 }
 
 type User = {
